chore(validation): implement create validation framework - move serializers folder into utilities and rename to model_serializers - add attribute model marshmallow schema - remove /register demo endpoint from manage module
[Maintains #157767356]
